requirements.txt
Flask==3.0.2
Pillow==10.2.0
qrcode==7.4.2
pyzbar==0.1.9 